coc: Disable snippets
Snippets are the kind of thing that I use a lot on IDEs but when
I'm on vim it feels more like it gets in the way.

The expansions, however, are super useful. Something like doing:

    var.box<tab>

And getting expanded to:

    Box::new(var)

And similar are very handy, especially the more complex ones
doing loops, if-lets, and stuff. I'd like to benefit from them, but
for now I'll take less features.