Drop coc.nvim, use more ALE
Turns out that the reason ALE's language server features weren't working
well was because I was trying it [before this rust-analyzer commit
went in][1]. Now it works.

Coc.nvim is still vastly superior in terms of features, but all that I
really care about are the `gd`, `K` mappings and the lsp completions.
And ALE doesn't require node.

[1]: https://github.com/rust-analyzer/rust-analyzer/commit/e075e096cf4970014d2c0829476fd7a45a3f32b1