Add a :GFilesRelative command
The original call just injecting `fnamemodify` would only work
correctly when the vim instance was running from the root of
the git repository.

This patch creates the `:GFilesRelative` command which uses
fugitive to find out the root of the repository then limits
it to the current path to the file of the active buffer. In
case it fails to find a git root, we simply fall back to
`:Files`.